requirements.txt
flask==3.0.3
python-dotenv==1.0.1
google-generativeai==0.7.2
openai==1.40.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
werkzeug==3.0.3
